requirements.txt
python-telegram-bot
Flask
gunicorn
requests
uvicorn
asgiref
flask[async]
asyncio